+++ Userland/DevTools/UserspaceEmulator/RangeAllocator.h
@@ -14,7 +14,6 @@ namespace UserspaceEmulator {
 class RangeAllocator {
 public:
     RangeAllocator();
-    ~RangeAllocator();
 
     void initialize_with_range(VirtualAddress, size_t);
